feat(node): isolate node dependants into a separate folder
