+++ packages/core/src/InfluxDB.ts
@@ -3,7 +3,7 @@ import {ClientOptions, WritePrecision} from './options'
 import WriteApiImpl from './impl/WriteApiImpl'
 import {IllegalArgumentError} from './errors'
 import {Transport} from './transport'
-import NodeHttpTransport from './impl/NodeHttpTransport'
+import NodeHttpTransport from './impl/node/NodeHttpTransport'
 import QueryApi from './QueryApi'
 import QueryApiImpl from './impl/QueryApiImpl'
 import SetupApi from './SetupApi'
